fix(ingredient): IngredientDto currentUnitPrice 타입 Int→Double 수정
서버가 currentUnitPrice를 소수점 형태(0.00, 30.00)로 반환하여
kotlinx.serialization 역직렬화 실패로 재료 목록이 표시되지 않던 버그 수정

+++ core/core-network/src/main/kotlin/com/team/chord/core/network/dto/ingredient/IngredientDtos.kt
@@ -16,7 +16,7 @@ data class IngredientDto(
     val ingredientName: String,
     val unitCode: String,
     val baseQuantity: Int,
-    val currentUnitPrice: Int,
+    val currentUnitPrice: Double,
 )
 
 @Serializable
